fix(backup-weight-threshold): give the unit selector a real 64-bit draw
The previous commit widened the weight totals to 64-bit and combined two
draws, but both halves were still derived from a single 32-bit
`random_seed`, so the draw carried only 32 bits of entropy: it could not
address the whole range once a hostgroup's total weight exceeded 2^32.

Worse, the 32-bit LCG multiplier made it worse than a mere entropy limit.
`draw = (state1 << 32) | (state2 & 0xffffffff)` keeps the seed's high bits
out of the draw entirely: a 32-bit multiplier shifted left by 32 cancels
the 2^32 offset, so seeds `i` and `i + 2^32` produced byte-identical draws
and therefore identical selections. The new test caught exactly that.

- `random_seed` is now `uint64_t` in the public helper and in both
  definitions, so all 64 bits reach the mixer.
- The 32-bit LCG is replaced by the splitmix64 finalizer, which mixes every
  input bit into the output. Existing distribution and determinism
  assertions are unchanged (50% / 75% / same-seed-same-result).
- New test sweeps seeds 0..199 and 2^32..2^32+199 over two 2^32-weight
  primaries and requires the two sequences to differ. It fails against the
  previous LCG and against a 32-bit seed parameter.

Spec: the 64-bit draw note said a 2^32-weight server became unreachable,
which is backwards - the first cumulative interval is the one a 32-bit draw
always lands in, so it is the *later* candidates that become unreachable.

+++ include/ServerSelection.h
@@ -79,7 +79,9 @@ bool is_candidate_eligible(const ServerCandidate &candidate);
  *
  * @param candidates Array of server candidates.
  * @param count      Number of candidates in the array.
- * @param random_seed Seed for deterministic random selection.
+ * @param random_seed Seed for deterministic random selection. 64-bit: the
+ *        weight totals are 64-bit, and a 32-bit seed could not address the
+ *        whole range once a hostgroup's total weight exceeds 2^32.
  * @param backup_weight_threshold T. T = 0 preserves today's behaviour.
  *        weight = 0 is never selected.
  * @param backup_availability How a primary is considered present. Pass 2
@@ -89,7 +91,7 @@ bool is_candidate_eligible(const ServerCandidate &candidate);
 int select_server_from_candidates(
 	const ServerCandidate *candidates,
 	int count,
-	unsigned int random_seed,
+	uint64_t random_seed,
 	int64_t backup_weight_threshold = 0,
 	BackupAvailability backup_availability = BACKUP_AVAIL_SELECTABLE
 );
